Added Home URLs
error-checking code now adds "url-home" to each feed's set of data. This string should hold the url that the rss feed represents (a blog homepage, ect). This will be used later (hopefully) to support links within output. I may add the ability to automagically pull the home url from the parsed feed in the same way titles are - I need to check that that is possible.